refactor: 디렉토리 네이밍 및 구조 수정
- controller -> presentation
- repository Class -> domain 내로 이동
- 테스트 코드도 동일하게 적용
