+++ backend/src/main/java/kr/kro/colla/project/project/domain/ProjectRepository.java
@@ -1,6 +1,5 @@
-package kr.kro.colla.project.project.repository;
+package kr.kro.colla.project.project.domain;
 
-import kr.kro.colla.project.project.domain.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProjectRepository extends JpaRepository<Project, Long> {
