+++ backend/src/main/java/kr/kro/colla/project/project/service/ProjectService.java
@@ -1,8 +1,8 @@
 package kr.kro.colla.project.project.service;
 
 import kr.kro.colla.project.project.domain.Project;
-import kr.kro.colla.project.project.repository.ProjectRepository;
-import kr.kro.colla.project.project.service.dto.CreateProjectRequest;
+import kr.kro.colla.project.project.domain.ProjectRepository;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
