+++ backend/src/test/java/kr/kro/colla/project/project/domain/ProjectRepositoryTest.java
@@ -1,6 +1,5 @@
-package kr.kro.colla.project.project.repository;
+package kr.kro.colla.project.project.domain;
 
-import kr.kro.colla.project.project.domain.Project;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
@@ -8,7 +7,8 @@ import org.springframework.test.context.ActiveProfiles;
 
 import javax.validation.ConstraintViolationException;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @ActiveProfiles("test")
 @DataJpaTest
